requirements.txt
css-html-js-minify~=2.5
docker~=5.0
GitPython~=3.1
Jinja2~=3.0
jsmin~=2.2
python-slugify~=5.0
six~=1.16
